added a test in `elementRemoved` to check if an element is in fact draggable/droppable before running the code to de-register it.

+++ src/katavorio.js
@@ -1262,8 +1262,12 @@
         };
 
         this.elementRemoved = function(el) {
-            this.destroyDraggable(el);
-            this.destroyDroppable(el);
+            if (el["_katavorioDrag"]) {
+                this.destroyDraggable(el);
+            }
+            if (el["_katavorioDrop"]) {
+                this.destroyDroppable(el);
+            }
         };
 
         /**
